+++ Scenes/Shared_Components/HurtBox.cs
@@ -6,7 +6,7 @@ public partial class HurtBox : Area2D
     [Export]
     public HealthComponent Health;
 
-    public void ReceiveAttack(double damage)
+    public void ReceiveAttack(int damage)
     {
         Health.TakeDamage(damage);
     }
